feat: implement auth API with login, logout, and token refresh functionality
