+++ client/src/App.jsx
@@ -38,6 +38,7 @@ function App() {
       locale={viVN}
     >
       <Router>
+        
         <Routes>
           <Route path="/login" element={<Login />} />
